requirements.txt
cachetools
numpy
flask
psycopg2-binary
obspy
pandas
pymysql
plotly
kaleido
pillow
ujson
python-dateutil
pytz
requests
pygmt
flask-compress
flask-session
pyyaml
